+++ kibot/__main__.py
@@ -12,9 +12,10 @@ Usage:
          [-q | -v...] [-L LOGFILE] [-C | -i | -n] [-m MKFILE] [-A] [-g DEF] ...
          [-E DEF] ... [-w LIST] [--banner N] [TARGET...]
   kibot [-v...] [-b BOARD] [-e SCHEMA] [-c PLOT_CONFIG] [--banner N]
-         [-E DEF] [--config-outs] [--only-pre] [--only-names] ... --list
-  kibot [-v...] [-c PLOT_CONFIG] [--banner N] [-E DEF] [--only-names] ...
-        --list-variants
+         [-E DEF] ... [--config-outs] [--only-pre|--only-groups] [--only-names]
+         --list
+  kibot [-v...] [-c PLOT_CONFIG] [--banner N] [-E DEF] ... [--only-names]
+         --list-variants
   kibot [-v...] [-b BOARD] [-d OUT_DIR] [-p | -P] [--banner N] --example
   kibot [-v...] [--start PATH] [-d OUT_DIR] [--dry] [--banner N]
          [-t, --type TYPE]... --quick-start
@@ -46,7 +47,8 @@ Options:
   -E DEF, --define DEF             Define preprocessor value (VAR=VAL)
   -g DEF, --global-redef DEF       Overwrite a global value (VAR=VAL)
   -i, --invert-sel                 Generate the outputs not listed as targets
-  -l, --list                       List available outputs (in the config file).
+  -l, --list                       List available outputs, preflights and
+                                   groups (in the config file).
                                    You don't need to specify an SCH/PCB unless
                                    using --config-outs
   --list-variants                  List the available variants and exit
@@ -55,9 +57,11 @@ Options:
   -m MKFILE, --makefile MKFILE     Generate a Makefile (no targets created)
   -n, --no-priority                Don't sort targets by priority
   -p, --copy-options               Copy plot options from the PCB file
-  --only-names                     Print only the names
-  --only-pre                       Print only the preflights. Note that
-                                   `--only-names` alone prints only the outputs
+  --only-names                     Print only the names. Note that for --list
+                                   if no other --only-* option is provided it
+                                   also acts as a virtual --only-outputs
+  --only-groups                    Print only the groups.
+  --only-pre                       Print only the preflights
   -P, --copy-and-expand            As -p but expand the list of layers
   -q, --quiet                      Remove information logs
   -s PRE, --skip-pre PRE           Skip preflights, comma separated or `all`
@@ -127,7 +131,7 @@ from .registrable import RegOutput
 GS.kibot_version = __version__
 
 
-def list_pre_and_outs_names(logger, outputs, do_config, only_pre):
+def list_pre_and_outs_names(logger, outputs, do_config, only_pre, only_groups):
     pf = BasePreFlight.get_in_use_names()
     if len(pf) and only_pre:
         for c in pf:
@@ -140,17 +144,17 @@ def list_pre_and_outs_names(logger, outputs, do_config, only_pre):
             logger.info(o.name)
 
 
-def list_pre_and_outs(logger, outputs, do_config, only_names, only_pre):
+def list_pre_and_outs(logger, outputs, do_config, only_names, only_pre, only_groups):
     if only_names:
-        return list_pre_and_outs_names(logger, outputs, do_config, only_pre)
+        return list_pre_and_outs_names(logger, outputs, do_config, only_pre, only_groups)
     pf = BasePreFlight.get_in_use_objs()
-    if len(pf):
+    groups = RegOutput.get_groups()
+    if pf and not only_groups:
         logger.info('Available pre-flights:')
         for c in pf:
             logger.info('- '+str(c))
-        if len(outputs) and not only_pre:
-            logger.info("")
-    if len(outputs) and not only_pre:
+        logger.info("")
+    if outputs and not only_pre and not only_groups:
         logger.info("Available outputs: 'comment/description' (name) [type]")
         for o in outputs:
             # Note: we can't do a `dry` config because some layer and field names can be validated only if we
@@ -158,14 +162,23 @@ def list_pre_and_outs(logger, outputs, do_config, only_names, only_pre):
             if do_config:
                 config_output(o, dry=False)
             logger.info('- '+str(o))
-    if len(pf):
         logger.info("")
+    if groups and not only_pre:
+        logger.info("Available groups:")
+        for g, items in groups.items():
+            logger.info('- '+g+': '+', '.join(items))
+        logger.info("")
+    if pf:
         logger.info("You can use e.g. `kibot --skip-pre preflight_name1,preflight_name2` to")
         logger.info("skip specific preflights (or pass `all` to skip them all).")
-    if len(outputs):
         logger.info("")
-        logger.info("You can use e.g. `kibot output_name1 output_name1` to generate only")
+    if outputs:
+        logger.info("You can use e.g. `kibot output_name1 output_name2` to generate only")
         logger.info("specific outputs by name.")
+        logger.info("")
+    if groups:
+        logger.info("You can use the name of a group instead of an output name.")
+        logger.info("")
 
 
 def list_variants(logger, only_names):
@@ -482,7 +495,7 @@ def main():
 
     # Is just "list the available targets"?
     if args.list:
-        list_pre_and_outs(logger, outputs, args.config_outs, args.only_names, args.only_pre)
+        list_pre_and_outs(logger, outputs, args.config_outs, args.only_names, args.only_pre, args.only_groups)
         sys.exit(0)
 
     if args.list_variants: