+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - General:
   - OS environment expansion in ${VAR}
+  - Now outputs can request to be added to one or more groups (#435)
 - Command line:
   - `--list-variants` List all available variants (See #434)
   - `--only-names` to make `--list` list only output names
   - `--only-pre` to list only the preflights
+  - `--only-groups` to list only the groups
 - Global options:
   - `use_os_env_for_expand` to disable OS environment expansion
   - `environment`.`extra_os` to define environment variables
@@ -33,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `_value_split_replace` splits the Value field and replaces it
 
 ### Changed
+- Command line:
+  - `--list` also lists groups
 - KiCad v6/7 schematic:
   - The hierarchy is expanded only if needed, i.e. value of an instance changed
 - List actions: